refactor(TextInput): move TextInput component to Generic directory
Consolidate the TextInput component into the Generic directory to improve maintainability and align with the project's component structure. This change involves moving the component and updating all import paths across the codebase.
