+++ src/components/Composite/admin/semesters/EditSemesterFormDialog.tsx
@@ -13,7 +13,7 @@ import {
   HStack,
   useNotice,
 } from '@yamada-ui/react'
-import { TextInput } from '../../TextInput'
+import { TextInput } from '../../../Generic/TextInput'
 import { useSemesterContext } from './SemestersContext'
 import { compareDate, formatDateInISO, validateDate } from './utils'
