+++ src/components/Composite/admin/view-sessions/EditGameSessionFormDialog.tsx
@@ -17,7 +17,7 @@ import {
   HStack,
   useNotice,
 } from '@yamada-ui/react'
-import { TextInput } from '../../TextInput'
+import { TextInput } from '../../../Generic/TextInput'
 import { useGameSessionContext } from './GameSessionContext'
 import { gameSessionFormSchema } from './utils'
