+++ src/components/Composite/auth/ResetPasswordForm.tsx
@@ -6,11 +6,11 @@ import { useForm } from 'react-hook-form'
 import { z } from 'zod'
 
 import { useResetPasswordMutation } from '@/hooks/mutations/reset-password'
+import { TextInput } from '../../Generic/TextInput/TextInput'
 import { Button } from '../../Generic/ui/button'
 import { ToastAction } from '../../Generic/ui/toast'
 import { useToast } from '../../Generic/ui/use-toast'
 import { Card } from '../Card'
-import { TextInput } from '../TextInput'
 import { passwordSchema } from './formSchema'
 
 const formSchema = z
